Renamed AlgorandClientSender to AlgorandClientTransactionSender for consistency
