--- docs/code/modules/types_algorand_client_sender.md
@@ -1,25 +1,0 @@
-[@algorandfoundation/algokit-utils](../README.md) / types/algorand-client-sender
-
-# Module: types/algorand-client-sender
-
-## Table of contents
-
-### Classes
-
-- [AlgorandClientSender](../classes/types_algorand_client_sender.AlgorandClientSender.md)
-
-### Type Aliases
-
-- [SendSingleTransactionResult](types_algorand_client_sender.md#sendsingletransactionresult)
-
-## Type Aliases
-
-### SendSingleTransactionResult
-
-Ƭ **SendSingleTransactionResult**: [`SendAtomicTransactionComposerResults`](../interfaces/types_transaction.SendAtomicTransactionComposerResults.md) & [`ConfirmedTransactionResult`](../interfaces/types_transaction.ConfirmedTransactionResult.md)
-
-Result from sending a single transaction.
-
-#### Defined in
-
-[src/types/algorand-client-sender.ts:10](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/algorand-client-sender.ts#L10)
