+++ docs/code/modules/types_algorand_client_transaction_sender.md
@@ -1,0 +1,25 @@
+[@algorandfoundation/algokit-utils](../README.md) / types/algorand-client-transaction-sender
+
+# Module: types/algorand-client-transaction-sender
+
+## Table of contents
+
+### Classes
+
+- [AlgorandClientTransactionSender](../classes/types_algorand_client_transaction_sender.AlgorandClientTransactionSender.md)
+
+### Type Aliases
+
+- [SendSingleTransactionResult](types_algorand_client_transaction_sender.md#sendsingletransactionresult)
+
+## Type Aliases
+
+### SendSingleTransactionResult
+
+Ƭ **SendSingleTransactionResult**: [`SendAtomicTransactionComposerResults`](../interfaces/types_transaction.SendAtomicTransactionComposerResults.md) & [`ConfirmedTransactionResult`](../interfaces/types_transaction.ConfirmedTransactionResult.md)
+
+Result from sending a single transaction.
+
+#### Defined in
+
+[src/types/algorand-client-transaction-sender.ts:10](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/algorand-client-transaction-sender.ts#L10)
