+++ src/types/account-manager.ts
@@ -2,7 +2,7 @@ import algosdk from 'algosdk'
 import { Config } from '../config'
 import { calculateFundAmount, memoize } from '../util'
 import { AccountInformation, DISPENSER_ACCOUNT, MultisigAccount, SigningAccount, TransactionSignerAccount } from './account'
-import { SendSingleTransactionResult } from './algorand-client-sender'
+import { SendSingleTransactionResult } from './algorand-client-transaction-sender'
 import { AlgoAmount } from './amount'
 import { ClientManager } from './client-manager'
 import AlgoKitComposer, { CommonTransactionParams, ExecuteParams } from './composer'
